+++ packages/expo-manifests/android/src/test/java/expo/modules/manifests/core/ManifestTest.kt
@@ -9,28 +9,35 @@ import org.junit.Test
 class ManifestTest {
   @Test
   fun getPluginProperties_emptyManifest_returnsNull() {
-    val manifestJson = JSONObject(emptyMap<String, Any>())
+    val manifestJson = JSONObject("{}")
     val manifest = Manifest.fromManifestJson(manifestJson)
     Truth.assertThat(manifest.getPluginProperties("test")).isNull()
   }
 
   @Test
   fun getPluginProperties_emptyPlugins_returnsNull() {
-    val manifestJson = JSONObject(mapOf<String, Any>("plugins" to emptyArray<Any>()))
+    val manifestJson = JSONObject("{\"plugins\": []}")
     val manifest = Manifest.fromManifestJson(manifestJson)
     Truth.assertThat(manifest.getPluginProperties("test")).isNull()
   }
 
   @Test
   fun getPluginProperties_nonMatchedPlugins_returnsNull() {
-    val manifestJson = JSONObject(mapOf<String, Any>("plugins" to arrayOf("hello")))
+    val manifestJson = JSONObject("{\"plugins\": [\"hello\"]}")
     val manifest = Manifest.fromManifestJson(manifestJson)
     Truth.assertThat(manifest.getPluginProperties("test")).isNull()
   }
 
   @Test
   fun getPluginProperties_matchedPluginWithoutProps_returnsNull() {
-    val manifestJson = JSONObject(mapOf<String, Any>("plugins" to arrayOf("test")))
+    val manifestJson = JSONObject("{\"plugins\": [\"test\"]}")
+    val manifest = Manifest.fromManifestJson(manifestJson)
+    Truth.assertThat(manifest.getPluginProperties("test")).isNull()
+  }
+
+  @Test
+  fun getPluginProperties_matchedPluginWithoutPropsAsNestedArray_returnsNull() {
+    val manifestJson = JSONObject("{\"plugins\": [[\"test\"]]}")
     val manifest = Manifest.fromManifestJson(manifestJson)
     Truth.assertThat(manifest.getPluginProperties("test")).isNull()
   }
@@ -38,8 +45,7 @@ class ManifestTest {
   @Test
   fun getPluginProperties_matchedPluginWithProps_returnsProps() {
     val props = mapOf<String, Any>("foo" to "bar")
-    val pluginWithProp = arrayOf("test", props)
-    val manifestJson = JSONObject(mapOf<String, Any>("plugins" to arrayOf(pluginWithProp)))
+    val manifestJson = JSONObject("{\"plugins\": [ [\"test\", {\"foo\":\"bar\"}] ]}")
     val manifest = Manifest.fromManifestJson(manifestJson)
     val result = manifest.getPluginProperties("test")
     Truth.assertThat(result).isNotNull()
@@ -49,8 +55,7 @@ class ManifestTest {
   @Test
   fun getPluginProperties_matchedPluginWithNestedProps_returnsNestedProps() {
     val props = mapOf<String, Any>("nested" to mapOf<String, Any>("insideNested" to true))
-    val pluginWithProp = arrayOf("test", props)
-    val manifestJson = JSONObject(mapOf<String, Any>("plugins" to arrayOf(pluginWithProp)))
+    val manifestJson = JSONObject("{\"plugins\":[[\"test\",{\"nested\":{\"insideNested\":true}}]]}")
     val manifest = Manifest.fromManifestJson(manifestJson)
     val result = manifest.getPluginProperties("test")
     Truth.assertThat(result).isNotNull()